Add split_rng=False (current default) to HTM.
Haiku plans to make split_rng a required argument to hk.vmap in an upcoming
release. This change updates HTM to preserve the current behaviour. We
also handle the case where users are using a release of Haiku without the
split_rng option, for these users split_rng=False is implied.

PiperOrigin-RevId: 428454975